+++ tests/test_volume.py
@@ -13,20 +13,20 @@ def volume_instance():
 
 
 def test_gain(volume_instance):
-    volume_instance._setting = 0.1
-    assert volume_instance.gain() == pytest.approx(1.03514216668)
-    volume_instance._setting = 0.5
-    assert volume_instance.gain() == pytest.approx(1.18850222744)
-    volume_instance._setting = 0.01
-    assert volume_instance.gain() == pytest.approx(0)
-    volume_instance._setting = 9.0
-    assert volume_instance.gain() == pytest.approx(22.3872113857)
-    volume_instance._setting = 0.01
+    volume_instance._volume_level = 0.1
+    assert volume_instance.calculate_gain() == pytest.approx(0.0462381021)
+    volume_instance._volume_level = 0.5
+    assert volume_instance.calculate_gain() == pytest.approx(0.0530884444)
+    volume_instance._volume_level = 0.01
+    assert volume_instance.calculate_gain() == pytest.approx(0)
+    volume_instance._volume_level = 9.0
+    assert volume_instance.calculate_gain() == pytest.approx(1.0)
+    volume_instance._volume_level = 0.01
     volume_instance._offset = 3.0
-    assert volume_instance.gain() == pytest.approx(0)
+    assert volume_instance.calculate_gain() == pytest.approx(0)
 
 
 def test_volume_initialization(volume_instance):
-    assert volume_instance._setting == 9
-    assert volume_instance._offset == 0
-    assert volume_instance._volume == pytest.approx(22.3872113857)
+    assert volume_instance._volume_level == 9
+    assert volume_instance._offset == 9
+    assert volume_instance._gain == pytest.approx(1.0)