+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.10] - 2023-11-14
+### Changed
+- Fixed existing unit tests for ADSR, Oscillator, and Volume classes
+### Removed
+- Removed unused Test_git_1.py file
+
 ## [1.1.9] - 2023-11-10
 ### Changed
 - Updated the readme